Increase version number and update copyright year.
I have used libalpm version 5.0.0 because that is the version
alpm_version returns.  I am not sure if I should use the libalpm
version (which probably no one knows/cares about) or the pacman
version (which would be easier to understand).

In my opinion, I would put the pacman version in the version
information for clyde.  At least until pacman and libalpm are
separated properly.