Allow manpage generation to fail.
git sets the file timestampts to the current time. If the clock changes
the second after the .ronn file is downloaded and before the .8 roff file
is downloaded make will believe the manpage is out of date and try to
generate it. This is bad when we are deploying with git to users who
do not have ronn installed.

Reported by Otto.