Rewrite umask() to avoid using tonumber(). "Better" umask() usage.
umask() in utilcore now takes a string as argument and does the octal
conversion itself. This means it also returns a string.

Changed clyde to set the umask right when starting up. Still some
problems with it. Not sure why but the previous umask is zero
when building AUR packages?

Code in aur.lua that changes the umask now restores it afterwords.